refactor(types): restructure utility types into grouped files
- Rename `array-types.ts` to  `arrays.ts`
- `deep.ts`
- Rename `object-types.ts` to `objects.ts`
- `strings.ts`
- `numbers.ts`
- Rename `type-guards` to `guards.ts`
- `utils.ts`
- Delete `types.ts`
