+++ src/index.ts
@@ -1,5 +1,8 @@
-export type * from "./utility-types.js"
-export type * from "./types.js"
+export type * from "./arrays.js"
+export type * from "./deep.js"
+export type * from "./guards.js"
+export type * from "./numbers.js"
+export type * from "./objects.js"
+export type * from "./strings.js"
 export type * from "./test.js"
-export type * from "./type-guards.js"
-export type * from "./string-mappers.js"
+export type * from "./utils.js"
