+++ src/utility-types.ts
@@ -1,3 +1,9 @@
-export type * from "./utils.js"
-export type * from "./array-types.js"
-export type * from "./object-types.js"
+/**
+ * @deprecated
+ * This file is deprecated and will be removed in the next major release.
+ * Please use the `@halvaradop/ts-utility-types` default export instead
+ * or see the others exports for specific utility types.
+ */
+throw new Error(
+    "[DEPRECATED] utility-types entry point is deprecated and will be removed in the next major release. Please use the default export from `@halvaradop/ts-utility-types` instead. Or use the specific utility types you need from the package directly."
+)
